+++ game.server/src/handlers/request/card.select.request.handler.ts
@@ -1,4 +1,4 @@
-import { GameSocket } from "../../utils/game.socket.js";
+import { GameSocket } from "../../type/game.socket.js";
 import { C2SCardSelectRequest } from "../../generated/packet/game_actions.js";
 import { GamePacket } from "../../generated/gamePacket.js";
 import { GamePacketType } from "../../enums/gamePacketType.js";
