+++ game.server/src/handlers/request/destroy.card.request.handler.ts
@@ -1,4 +1,4 @@
-import { GameSocket } from "../../utils/game.socket.js";
+import { GameSocket } from "../../type/game.socket.js";
 import { C2SDestroyCardRequest } from "../../generated/packet/game_actions.js";
 import { GamePacket } from "../../generated/gamePacket.js";
