+++ game.server/src/handlers/request/join.random.room.request.handler.ts
@@ -1,4 +1,4 @@
-import { GameSocket } from "../../utils/game.socket.js";
+import { GameSocket } from "../../type/game.socket.js";
 import { C2SJoinRandomRoomRequest } from "../../generated/packet/room_actions.js";
 import { GamePacket } from "../../generated/gamePacket.js";
