+++ game.server/src/handlers/response/join.random.room.response.handler.ts
@@ -1,4 +1,4 @@
-import { GameSocket } from "../../utils/game.socket.js";
+import { GameSocket } from "../../type/game.socket.js";
 import { S2CJoinRandomRoomResponse } from "../../generated/packet/room_actions.js";
 import { GamePacket } from "../../generated/gamePacket.js";
