+++ game.server/src/handlers/response/leave.room.response.handler.ts
@@ -1,4 +1,4 @@
-import { GameSocket } from "../../utils/game.socket.js";
+import { GameSocket } from "../../type/game.socket.js";
 import { S2CLeaveRoomResponse } from "../../generated/packet/room_actions.js";
 import { GamePacket } from "../../generated/gamePacket.js";
