+++ game.server/src/handlers/response/register.response.handler.ts
@@ -1,10 +1,10 @@
-import { GameSocket } from "../../utils/game.socket.js";
+import { GameSocket } from "../../type/game.socket.js";
 import { S2CRegisterResponse } from "../../generated/packet/auth.js";
 import { GamePacket } from "../../generated/gamePacket.js";
 import { getGamePacketType } from "../../utils/type.converter.js";
 import { GamePacketType, gamePackTypeSelect } from "../../enums/gamePacketType.js";
 import { GlobalFailCode } from "../../generated/common/enums.js";
-import { sendData } from "../../utils/sendData.js";
+import { sendData } from "../../utils/send.data.js";
 
 const registerResponseHandler = (socket: GameSocket, gamePacket: GamePacket) => {
