+++ game.server/src/type/game.socket.d.ts
@@ -2,4 +2,5 @@ import { Socket } from "net";
 
 export interface GameSocket extends Socket{
     userId?:string;
+    roomId?:number;
 }
